refactor: adjusting the structure for the tests
