+++ .gitignore
@@ -2,3 +2,4 @@
 .DS_Store
 .vscode/
 .idea
+coverage.out
