+++ internal/repository/buyer/buyer.go
@@ -41,6 +41,28 @@ func (r *buyerRepository) Create(ctx context.Context, b models.Buyer) (*models.B
 	return &b, nil
 }
 
+// func (r *buyerRepository) Delete(ctx context.Context, id int) error {
+// 	result, err := r.mysql.ExecContext(ctx, queryBuyerDelete, id)
+// 	if err != nil {
+// 		if mysqlErr, ok := err.(*mysql.MySQLError); ok && mysqlErr.Number == 1451 {
+// 			return apperrors.NewAppError(apperrors.CodeConflict, "Cannot delete buyer: there are purchase orders associated with this buyer.")
+// 		}
+// 		return apperrors.NewAppError(apperrors.CodeInternal,
+// 			fmt.Sprintf("An internal server error occurred while deleting the buyer: %s", err.Error()))
+// 	}
+
+// 	rowsAffected, err := result.RowsAffected()
+// 	if err != nil {
+// 		return apperrors.NewAppError(apperrors.CodeInternal,
+// 			fmt.Sprintf("An internal server error occurred while deleting the buyer: %s", err.Error()))
+// 	}
+
+//		if rowsAffected == 0 {
+//			return apperrors.NewAppError(apperrors.CodeNotFound, "The buyer you are trying to delete does not exist")
+//		}
+//		return nil
+//	}
+
 func (r *buyerRepository) Update(ctx context.Context, id int, b models.Buyer) error {
 	_, err := r.mysql.ExecContext(
 		ctx,
@@ -55,24 +77,30 @@ func (r *buyerRepository) Update(ctx context.Context, id int, b models.Buyer) er
 	}
 	return nil
 }
-
 func (r *buyerRepository) Delete(ctx context.Context, id int) error {
 	result, err := r.mysql.ExecContext(ctx, queryBuyerDelete, id)
 	if err != nil {
-		if mysqlErr, ok := err.(*mysql.MySQLError); ok && mysqlErr.Number == 1451 {
-			return apperrors.NewAppError(apperrors.CodeConflict, "Cannot delete buyer: there are purchase orders associated with this buyer.")
+		if mysqlErr, ok := err.(*mysql.MySQLError); ok {
+			if mysqlErr.Number == 1451 {
+				return apperrors.NewAppError(apperrors.CodeConflict,
+					"cannot delete buyer: there are purchase orders associated")
+			}
 		}
-		return apperrors.NewAppError(apperrors.CodeInternal, fmt.Sprintf("An internal server error occurred while deleting the buyer: %s", err.Error()))
+		return apperrors.NewAppError(apperrors.CodeInternal,
+			fmt.Sprintf("an internal server error occurred while deleting the buyer: %s", err.Error()))
 	}
 
 	rowsAffected, err := result.RowsAffected()
 	if err != nil {
-		return apperrors.NewAppError(apperrors.CodeInternal, fmt.Sprintf("An internal server error occurred while deleting the buyer: %s", err.Error()))
+		return apperrors.NewAppError(apperrors.CodeInternal,
+			"error verifying affected rows")
 	}
 
 	if rowsAffected == 0 {
-		return apperrors.NewAppError(apperrors.CodeNotFound, "The buyer you are trying to delete does not exist")
+		return apperrors.NewAppError(apperrors.CodeNotFound,
+			"buyer not found")
 	}
+
 	return nil
 }
 
@@ -83,7 +111,8 @@ func (r *buyerRepository) FindAll(ctx context.Context) ([]models.Buyer, error) {
 	}
 	defer rows.Close()
 
-	var buyers []models.Buyer
+	//var buyers []models.Buyer
+	buyers := []models.Buyer{}
 	for rows.Next() {
 		var b models.Buyer
 		err := rows.Scan(&b.Id, &b.CardNumberId, &b.FirstName, &b.LastName)
