+++ internal/repository/buyer/buyer_interface.go
